# gabarita-ai-backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Gabarita AI API
  description: |
    API do sistema Gabarita AI para geração de questões, gerenciamento de usuários e planos.
    
    ## Autenticação
    A API utiliza autenticação via JWT tokens. Inclua o token no header:
    ```
    Authorization: Bearer <seu_token_jwt>
    ```
    
    ## Formato de Resposta
    Todas as respostas seguem o padrão:
    ```json
    {
      "success": true|false,
      "data": { /* dados da resposta */ },
      "message": "Mensagem descritiva",
      "error": "Tipo do erro (apenas em caso de erro)"
    }
    ```
  version: 1.0.0
  contact:
    name: Gabarita AI Support
    email: suporte@gabaritai.app.br
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://gabarita-ai-backend.onrender.com
    description: Servidor de Produção
  - url: http://localhost:5000
    description: Servidor de Desenvolvimento
paths:
...
  /api/user/{user_id}:
    get:
      tags:
        - Usuários
      summary: Obter usuário por ID
      description: Retorna dados de um usuário específico
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: ID do usuário
      responses:
        '200':
          description: Usuário encontrado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Usuários
      summary: Atualizar usuário
      description: Atualiza dados de um usuário específico
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Usuário atualizado com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
  /api/user/profile:
    get:
      tags:
        - Usuários
      summary: Obter perfil do usuário logado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil do usuário
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'

    put:
      tags:
        - Usuários
      summary: Atualizar perfil do usuário logado
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Perfil atualizado com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'

  # ALIASES DE USUÁRIOS
...
  /api/questoes/gerar:
    post:
      tags:
        - Questões
      summary: Gerar nova questão
      description: Gera uma questão personalizada baseada nos parâmetros fornecidos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cargo
                - bloco
                - conteudo
              properties:
                cargo:
                  type: string
                  description: Cargo para o qual gerar a questão
                bloco:
                  type: string
                  description: Bloco de conteúdo
                conteudo:
                  type: string
                  description: Conteúdo específico
                dificuldade:
                  type: string
                  enum: [facil, medio, dificil]
                  default: medio
      responses:
        '200':
          description: Questão gerada com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Question'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operação realizada com sucesso
      required:
        - success
        - message
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Erro de validação
        message:
          type: string
          example: Descrição detalhada do erro
      required:
        - success
        - error
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
          description: ID único do usuário
        email:
          type: string
          format: email
          description: Email do usuário
        name:
          type: string
          description: Nome completo do usuário
        created_at:
          type: string
          format: date-time
          description: Data de criação da conta
        updated_at:
          type: string
          format: date-time
          description: Data da última atualização
        plan:
          type: string
          nullable: true
          description: Plano atual do usuário
      required:
        - id
        - email
        - name
        - created_at
        - updated_at
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          description: Nome completo do usuário
        email:
          type: string
          format: email
          description: Email do usuário
        password:
          type: string
          format: password
          description: Nova senha (opcional)
    Question:
      type: object
      properties:
        id:
          type: string
          description: ID único da questão
        enunciado:
          type: string
          description: Texto da questão
        alternativas:
          type: object
          properties:
            A:
              type: string
            B:
              type: string
            C:
              type: string
            D:
              type: string
            E:
              type: string
          description: Alternativas da questão
        cargo:
          type: string
          description: Cargo relacionado
        bloco:
          type: string
          description: Bloco de conteúdo
        conteudo:
          type: string
          description: Conteúdo específico
        dificuldade:
          type: string
          enum: [facil, medio, dificil]
          description: Nível de dificuldade
      required:
        - id
        - enunciado
        - alternativas
        - cargo
        - bloco
        - conteudo
        - dificuldade
...